Removed ItemRepository class and added GitHub Action
The ItemRepository class was removed from item_routes.py. This change simplifies the code by removing unnecessary classes. In addition, a new GitHub Action for Python application test was included in the .github/workflows directory, allowing for automatic testing of the Python application upon each push to the main branch. Also, a local SQLite data source was added to the .idea directory.