Add VPN and server deployment to GitHub workflow
This commit implements the steps required to set up a VPN connection within the GitHub Actions workflow. It also includes the deployment procedure to a specified server after the successful connection of VPN. Necessary secrets are used to access the VPN configuration and private SSH key for secure server access.