Refactor GitHub action workflows
This commit revamps the workflows for pylint and main, adjusting the process to use Python 3.10 and handle different actions. The pylint workflow will now run on both push and pull request events, and the main workflow is adjusted to accommodate VPN connection and SSH usage. The python-app workflow has also been updated to use GitHub secrets for server username and host details.